feat: robust policy_peek reshuffle and serialization for all players - Ensure deck reshuffle occurs automatically before policy_peek if fewer than 3 cards remain, for both human and bot presidents - Safely serialize top_policies in policy_peek to support both Policy objects and string types - Prevent JSON serialization errors when deck contains mixed or unexpected types - Unify policy_peek logic for bots and humans for consistent executive power behavior - Improve reliability of executive power endpoints and game flow
